Refactor EventsTable component to streamline event details presentation
- Removed the TimingWaterfall component to simplify the event details view, enhancing clarity and focus on essential information.
- Introduced EventDetailsRow for improved organization of event data, including status, error messages, and response headers.
- Updated loading skeletons for better visual consistency during data fetching.
- Adjusted timestamp display for a cleaner presentation of event times.
